ab_riscv_benchmarks::ffi

Module ed25519_verify

Structs

BenchmarksEd25519VerifyArgs

Data structure containing expected input for external method invocation, eventually calling benchmarks_ed25519_verify() on the other side by the host.

InternalArgs

Data structure containing expected input to benchmarks_ed25519_verify(), it is used internally by the contract, there should be no need to construct it explicitly except maybe in contract’s own tests

Constants

METADATA

Method metadata, see ContractMetadataKind for encoding details

Functions

benchmarks_ed25519_verify⚠

FFI interface into a method, called by the host.